// appliance-mgmt-ui/src/app/models/appliance-details-payload.ts
export class ApplianceDetailsPayload{
    "id": number;
    "applianceName": string;
    "applianceModel": string;
    "location": string;
    "generation": string;
    "canBeShared": boolean;
    "appliancePossessionStatus": string;
    "applianceState": string;
    "uomName": string;
    "purpose":string;
    "assignee": string;
    "assigneeEmail": string;
    "configuration": string;
    "startDate": string;
    "endDate": string;
}